dataset/python-mutated/otp.py
import itertools
import os
import time
from cryptography.hazmat.backends import default_backend
from cryptography.hazmat.primitives.hashes import SHA1
from cryptography.hazmat.primitives.twofactor import InvalidToken
from cryptography.hazmat.primitives.twofactor.totp import TOTP
TOTP_LENGTH = 6
TOTP_INTERVAL = 30

class OutOfSyncTOTPError(Exception):
    pass

class InvalidTOTPError(Exception):
    pass

def _get_totp(secret):
    if False:
        while True:
            i = 10
    '\n    Returns a TOTP object for device provisioning and OTP validation.\n\n    The TOTP object is instantiated with the default OTP parameters,\n    per RFC6238:\n        * SHA1 digest\n        * 6-digit code\n        * 30-second interval\n    '
    return TOTP(secret, TOTP_LENGTH, SHA1(), TOTP_INTERVAL, backend=default_backend())

def generate_totp_secret():
    if False:
        return 10
    '\n    Generates a secret for time-based OTP.\n\n    The default secret length is 160 bits, as per RFC4226:\n    https://tools.ietf.org/html/rfc4226#section-4\n    '
    return os.urandom(20)

def generate_totp_provisioning_uri(secret, username, issuer_name):
    if False:
        for i in range(10):
            print('nop')
    '\n    Generates a URL to be presented as a QR-code for time-based OTP.\n    '
    totp = _get_totp(secret)
    return totp.get_provisioning_uri(username, issuer_name)

def _verify_totp_time(totp, value, time):
    if False:
        return 10
    '\n    Verifies an OTP value and time against the given TOTP object.\n    '
    try:
        totp.verify(value, time)
        return True
    except InvalidToken:
        return False

def verify_totp(secret, value):
    if False:
        print('Hello World!')
    '\n    Verifies a given TOTP-secret and value for the\n    current time +/- 1 counter interval.\n\n    This minimizes issues caused by clock differences and latency,\n    provides a better UX, and also improves accessibility\n    in cases where typing speed is limited.\n    '
    totp = _get_totp(secret)
    now = time.time()
    valid_in_window = _verify_totp_time(totp, value, now) or _verify_totp_time(totp, value, now - TOTP_INTERVAL) or _verify_totp_time(totp, value, now + TOTP_INTERVAL)
    valid_outside_window = any((_verify_totp_time(totp, value, now + step) for step in itertools.chain(range(-TOTP_INTERVAL * 10, -TOTP_INTERVAL, TOTP_INTERVAL), range(TOTP_INTERVAL * 2, TOTP_INTERVAL * 11, TOTP_INTERVAL))))
    if valid_in_window:
        return True
    elif valid_outside_window:
        raise OutOfSyncTOTPError
    else:
        raise InvalidTOTPError